Refactor position, dimensions, and rectangle
