+++ src/components/shapes/Rectangle.tsx
@@ -1,20 +1,8 @@
 ﻿import { Position } from "./Position.ts";
+import { Dimensions } from "./Dimensions.ts";
 
 export interface Rectangles {
   [id: string]: Rectangle;
 }
 
-export interface Rectangle {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-export function getRectangle(start: Position, end: Position): Rectangle {
-  const x = Math.min(start.x, end.x);
-  const y = Math.min(start.y, end.y);
-  const width = Math.abs(start.x - end.x);
-  const height = Math.abs(start.y - end.y);
-  return { x, y, width, height };
-}
+export type Rectangle = Position & Dimensions;
